+++ lib/cocoa/src/transport/TNSFileHandleTransport.m
@@ -51,7 +51,7 @@
 }
 
 
-- (int) readAll: (uint8_t *) buf offset: (int) off length: (int) len
+- (size_t) readAll: (uint8_t *) buf offset: (size_t) off length: (size_t) len
 {
   int got = 0;
   while (got < len) {
@@ -67,7 +67,7 @@
 }
 
 
-- (void) write: (const uint8_t *) data offset: (unsigned int) offset length: (unsigned int) length
+- (void) write: (const uint8_t *) data offset: (size_t) offset length: (size_t) length
 {
   void *pos = (void *) data + offset;
   NSData * dataObject = [[NSData alloc] initWithBytesNoCopy: pos // data+offset
